requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.5
django-appconf==1.2.0
django-bootstrap5==25.2
django-ckeditor-5==0.2.18
django-imagekit==6.0.0
django-js-asset==3.1.2
idna==3.11
MarkupSafe==3.0.3
pilkit==3.0
pillow==11.3.0
pycparser==2.23
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
stripe==14.0.0
typing_extensions==4.15.0
urllib3==2.5.0
